Remove different ffmpeg path in dev
When I started working on this there was a different function to get the ffmpeg path, it was commented out so the code was crashing. By uncommenting, in a previous commit, it worked again, but I don't see why would it have to be a different function specifically from development, so with this commit the function when running in development is the same and it just works.