Encoder working and form is no longer empty. (#33)
* Use state to set encode method

The "ref" state for selecting the encoding method was invalid. By using the state we can properly set the encoding method as a string on its state and properly pass the string of the selected method.

* Use local ffmpegpath

It was throwing an error because there was no ffmpeg path in development mode, it was simply commented out, so I commented it back in and it can now encode in dev mode.

* Remove different ffmpeg path in dev

When I started working on this there was a different function to get the ffmpeg path, it was commented out so the code was crashing. By uncommenting, in a previous commit, it worked again, but I don't see why would it have to be a different function specifically from development, so with this commit the function when running in development is the same and it just works.

* change usage of ref to useState

* useState for each item of publishForm

* fix setState for title, decription and tags